Move pytorch_spdl to prototype directory
It's a prototype
