--- src/pytorch_spdl/__init__.py
@@ -1,1 +1,0 @@
-# pyre-strict
